fix: case sensitivity divider.tsx → Divider.tsx
