--- src/libs/divider/divider.module.css
@@ -1,10 +1,0 @@
-.divider {
-  border: none;
-  background-color: #007bff;
-  width: 100%;
-  height: 2px;
-}
-
-.divider-color--secondary {
-  background-color: #e0e0e0;
-}
